feat: BattleWebsocketHandler에 로깅 적용
