+++ src/main/java/online/partyrun/partyrunbattleservice/domain/battle/controller/BattleWebsocketController.java
@@ -6,6 +6,7 @@ import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 
+import lombok.extern.slf4j.Slf4j;
 import online.partyrun.partyrunbattleservice.domain.battle.dto.RunnerDistanceResponse;
 import online.partyrun.partyrunbattleservice.domain.battle.dto.RunnerRecordRequest;
 import online.partyrun.partyrunbattleservice.domain.battle.infra.RedisPublisher;
@@ -16,6 +17,7 @@ import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 
+@Slf4j
 @Controller
 @RequiredArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
@@ -35,5 +37,6 @@ public class BattleWebsocketController {
         final RunnerDistanceResponse response = battleService.calculateDistance(battleId, runnerId, request);
 
         messagePublisher.publish(response);
+        log.info(response.toString());
     }
 }
